+++ modules/tinymce/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Added support for alpha list numbering to the `lists` plugin #TINY-6891
+
 ## 5.7.0 - 2021-02-10
 
 ### Added